+++ lib/encodings.dart
@@ -28,29 +28,29 @@ class EncodingUtils {
   }
 
   static final _regexMap = {
-    Encoding.StandardRFC4648: RegExp(r'^[A-Z2-7=]+$'),
-    Encoding.Base32Hex: RegExp(r'^[0-9A-V=]+$'),
-    Encoding.Crockford: RegExp(r'^[0123456789ABCDEFGHJKMNPQRSTVWXYZ]+$'),
-    Encoding.ZBase32: RegExp(r'^[ybndrfg8ejkmcpqxot1uwisza345h769]+$'),
-    Encoding.Geohash: RegExp(r'^[0123456789bcdefghjkmnpqrstuvwxyz=]+$')
+    Encoding.standardRFC4648: RegExp(r'^[A-Z2-7=]+$'),
+    Encoding.base32Hex: RegExp(r'^[0-9A-V=]+$'),
+    Encoding.crockford: RegExp(r'^[0123456789ABCDEFGHJKMNPQRSTVWXYZ]+$'),
+    Encoding.zbase32: RegExp(r'^[ybndrfg8ejkmcpqxot1uwisza345h769]+$'),
+    Encoding.geohash: RegExp(r'^[0123456789bcdefghjkmnpqrstuvwxyz=]+$')
   };
   static final _encodeMap = {
-    Encoding.StandardRFC4648: 'ABCDEFGHIJKLMNOPQRSTUVWXYZ234567',
-    Encoding.Base32Hex: '0123456789ABCDEFGHIJKLMNOPQRSTUV',
-    Encoding.Crockford: '0123456789ABCDEFGHJKMNPQRSTVWXYZ',
-    Encoding.ZBase32: 'ybndrfg8ejkmcpqxot1uwisza345h769',
-    Encoding.Geohash: '0123456789bcdefghjkmnpqrstuvwxyz'
+    Encoding.standardRFC4648: 'ABCDEFGHIJKLMNOPQRSTUVWXYZ234567',
+    Encoding.base32Hex: '0123456789ABCDEFGHIJKLMNOPQRSTUV',
+    Encoding.crockford: '0123456789ABCDEFGHJKMNPQRSTVWXYZ',
+    Encoding.zbase32: 'ybndrfg8ejkmcpqxot1uwisza345h769',
+    Encoding.geohash: '0123456789bcdefghjkmnpqrstuvwxyz'
   };
 
   static final Map<Encoding, Map<String, int>> _decodeMap = {};
 
   static final _padded = {
-    Encoding.StandardRFC4648: true,
-    Encoding.Base32Hex: true,
-    Encoding.Crockford: false,
-    Encoding.ZBase32: false,
-    Encoding.Geohash: true
+    Encoding.standardRFC4648: true,
+    Encoding.base32Hex: true,
+    Encoding.crockford: false,
+    Encoding.zbase32: false,
+    Encoding.geohash: true
   };
 }
 
-enum Encoding { StandardRFC4648, Base32Hex, Crockford, ZBase32, Geohash }
+enum Encoding { standardRFC4648, base32Hex, crockford, zbase32, geohash }